feat: (Webhook Forwarder) Combine NX and Salient
